for combinedLoggedTest added logout, added delete account screen check, added sign-in as a before function, added settings screen test, added navigation item buttons check from the main screen. For combinedNotLogged added sending a massage to support on support screen, added main screen verification for navigation buttons. Added reserveFelyx test. Deleted all tests that were included in combined tests. Cleaning up commented lines that are not required  was done for all tests.
