+++ config/wdio.android.app.test.conf.ts
@@ -5,7 +5,7 @@ export const config: WebdriverIO.Config = {
     ...baseConfig,
 
     //specs: ["/Users/alesav/Dev/umob/appium-boilerplate/tests/specs/book/bookLime.spec.ts"],
-    specs: ["c:/dev/umob_appium/tests/specs/book/bookTaxi.spec.ts"],
+    specs: ["c:/dev/umob_appium/tests/specs/book/reserveFelyx.spec.ts"],
     capabilities: [
         {
             platformName: "Android",
