--- tests/specs/account/idDocument.spec.ts
@@ -1,93 +1,0 @@
-describe('ID Document Screen Elements Validation', () => {
-  beforeEach(async () => {
-    await driver.activateApp("com.umob.umob");
-    await driver.pause(7000);
-  });
-
-  it('should display all key ID Document screen elements', async () => {
-    // Navigate to Account screen first
-    const accountButton = await driver.$("-android uiautomator:new UiSelector().text(\"Account\")");
-    await accountButton.click();
-
-    // Navigate to ID Document screen
-    const idDocumentButton = await driver.$("-android uiautomator:new UiSelector().text(\"ID Document\")");
-    await expect(idDocumentButton).toBeDisplayed();
-    await idDocumentButton.click();
-
-    // Verify screen header
-    const screenHeader = await driver.$("-android uiautomator:new UiSelector().text(\"ID document\")");
-    await expect(screenHeader).toBeDisplayed();
-
-    // Verify back button is present
-    const backButton = await driver.$("-android uiautomator:new UiSelector().description(\"back_button\")");
-    await expect(backButton).toBeDisplayed();
-
-    // Verify "Verified" status text
-    const verifiedStatus = await driver.$("-android uiautomator:new UiSelector().text(\"Verified\")");
-    await expect(verifiedStatus).toBeDisplayed();
-
-    // Verify Document Type section
-    const documentTypeTitle = await driver.$("-android uiautomator:new UiSelector().text(\"Document type\")");
-    await expect(documentTypeTitle).toBeDisplayed();
-
-    const driverLicenseText = await driver.$("-android uiautomator:new UiSelector().text(\"Driver license\")");
-    await expect(driverLicenseText).toBeDisplayed();
-
-    // Verify Status section
-    const statusTitle = await driver.$("-android uiautomator:new UiSelector().text(\"Status\")");
-    await expect(statusTitle).toBeDisplayed();
-
-    const statusVerifiedText = await driver.$("-android uiautomator:new UiSelector().description(\"IdDocumentStatusIdDocumentItemContent\")");
-    await expect(statusVerifiedText).toBeDisplayed();
-
-    // Verify Expiration Date section
-    const expirationTitle = await driver.$("-android uiautomator:new UiSelector().text(\"Expiration date\")");
-    await expect(expirationTitle).toBeDisplayed();
-
-    const expirationDate = await driver.$("-android uiautomator:new UiSelector().text(\"28 May 2031\")");
-    await expect(expirationDate).toBeDisplayed();
-
-    // Verify Categories section
-    const categoriesTitle = await driver.$("-android uiautomator:new UiSelector().text(\"Categories\")");
-    await expect(categoriesTitle).toBeDisplayed();
-
-    // Verify all license categories
-    const categoryB = await driver.$("-android uiautomator:new UiSelector().description(\"undefinedIdDocumentItemContentB\")");
-    await expect(categoryB).toBeDisplayed();
-
-    const categoryA = await driver.$("-android uiautomator:new UiSelector().description(\"undefinedIdDocumentItemContentA\")");
-    await expect(categoryA).toBeDisplayed();
-
-    const categoryB1 = await driver.$("-android uiautomator:new UiSelector().description(\"undefinedIdDocumentItemContentB1\")");
-    await expect(categoryB1).toBeDisplayed();
-
-    const categoryAM = await driver.$("-android uiautomator:new UiSelector().description(\"undefinedIdDocumentItemContentAM\")");
-    await expect(categoryAM).toBeDisplayed();
-
-    // Verify Home Address section
-    const homeAddressTitle = await driver.$("-android uiautomator:new UiSelector().text(\"Home address\")");
-    await expect(homeAddressTitle).toBeDisplayed();
-
-    const addAddressButton = await driver.$("-android uiautomator:new UiSelector().text(\"ADD\")");
-    await expect(addAddressButton).toBeDisplayed();
-
-    // Verify bottom buttons
-    const changeDocumentButton = await driver.$("-android uiautomator:new UiSelector().text(\"CHANGE DOCUMENT\")");
-    await expect(changeDocumentButton).toBeDisplayed();
-
-    const helpButton = await driver.$("-android uiautomator:new UiSelector().text(\"Help\")");
-    await expect(helpButton).toBeDisplayed();
-
-    // Optional: Verify main container
-    const idDocumentContainer = await driver.$("-android uiautomator:new UiSelector().description(\"IdDocumentContainer\")");
-    await expect(idDocumentContainer).toBeDisplayed();
-  });
-
-  afterEach(async () => {
-    try {
-      await driver.terminateApp("com.umob.umob");
-    } catch (error) {
-      console.log('Error terminating app:', error);
-    }
-  });
-});
