--- tests/specs/account/inviteFriends.spec.ts
@@ -1,54 +1,0 @@
-describe('Invite Friends Screen Elements Validation', () => {
-  beforeEach(async () => {
-    await driver.activateApp("com.umob.umob");
-    await driver.pause(7000);
-  });
-
-  it('should display all key Invite Friends screen elements', async () => {
-    // Navigate to Account
-    const accountButton = await driver.$("-android uiautomator:new UiSelector().text(\"Account\")");
-    await accountButton.click();
-    await driver.pause(3000);
-
-    // Navigate to Invite Friends
-    const inviteFriendsButton = await driver.$("-android uiautomator:new UiSelector().text(\"Invite friends\")");
-    await inviteFriendsButton.click();
-    await driver.pause(3000);
-
-    // Verify back button is present
-    const backButton = await driver.$("-android uiautomator:new UiSelector().description(\"back_button\")");
-    await expect(backButton).toBeDisplayed();
-
-    // Verify screen title
-    const screenTitle = await driver.$("-android uiautomator:new UiSelector().text(\"Invite friends and earn €10 for each one!\")");
-    await expect(screenTitle).toBeDisplayed();
-
-    // Verify screen description
-    const screenDescription = await driver.$("-android uiautomator:new UiSelector().textContains(\"Make a friend ride with umob - both get €10,- ride credit\")");
-    await expect(screenDescription).toBeDisplayed();
-
-    // Verify Your Code section
-    const yourCodeLabel = await driver.$("-android uiautomator:new UiSelector().text(\"Your code\")");
-    await expect(yourCodeLabel).toBeDisplayed();
-
-    /* Verify the actual referral code
-    const referralCode = await driver.$("-android uiautomator:new UiSelector().text(\"QYI-S50\")");
-    await expect(referralCode).toBeDisplayed(); */
-
-    // Verify usage count
-    const usageCount = await driver.$("-android uiautomator:new UiSelector().text(\"Your code has been used 0 out of 5 times\")");
-    await expect(usageCount).toBeDisplayed();
-
-    // Verify Share Code button
-    const shareCodeButton = await driver.$("-android uiautomator:new UiSelector().text(\"SHARE CODE\")");
-    await expect(shareCodeButton).toBeDisplayed();
-  });
-
-  afterEach(async () => {
-    try {
-      await driver.terminateApp("com.umob.umob");
-    } catch (error) {
-      console.log('Error terminating app:', error);
-    }
-  });
-});
